Update USER_ConnectToOracleInstance.md
Update the "Example To connect to an Oracle DB instance using SQL*Plus" section to let the customer know they must modify the username to match their username of the database administrator for the DB instance. This will warn the customers that copy and paste the command, to update the username avoiding the command failing due to an unknown username.