Update Welcome.md
On line #11 the "or" should be replaced with an "and". The sentence is talking about tasks in the plural so "and" makes more sense. To illustrate, you might hire a sophisticated secretary to handle difficult and tedious tasks (plural). But this person would be capable of handling a single task (singular) that is  difficult or tedious. 

Thank you for considering my proposed change. 

-Brian Lange