Normalize file name
The "import the certificate" step was creating a keystore named `clientkeystore`, but then the following "confirm" step was trying to open a file named `clientkeystore.jks`. This caused a "Keystore file does not exist" error to be thrown, and correcting the commands to both reference the same file resolves the error.