Update to SSRS options for email subscriptions for SQL Server
The property originally given here `SMTP_SERVER_NAME` is incorrect. When deploying the email subscription setup for SSRS options it should be `SMTP_SERVER`. This was information I received from support when having problems with this property so figured I would PR the change to documentation so no one else hits the same issue.